# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Recipe & Meal Planner GraphQL API",
    "description": "GraphQL API for recipes and users, plus OAuth authentication flows.",
    "version": "1.0.0"
  },
  "host": "",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/graphql": {
      "post": {
        "summary": "Execute GraphQL operations",
        "description": "Submit a GraphQL query or mutation to the API. Use GraphiQL for interactive exploration in non-production environments.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "query": {
                  "type": "string",
                  "example": "query { recipes { _id title category difficulty } }"
                },
                "variables": {
                  "type": "object"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GraphQL response object."
          },
          "400": {
            "description": "Bad request or invalid GraphQL syntax."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/auth/google": {
      "get": {
        "summary": "Start Google OAuth login",
        "description": "Redirects the client to Google for OAuth authentication.",
        "responses": {
          "302": {
            "description": "Redirect to Google login."
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "summary": "Google OAuth callback",
        "description": "Handles Google OAuth callback and signs the user in.",
        "responses": {
          "200": {
            "description": "Login successful."
          },
          "401": {
            "description": "Authentication failed."
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Logout current user",
        "description": "Ends the authenticated session.",
        "responses": {
          "200": {
            "description": "Logout successful."
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get current user",
        "description": "Returns the currently authenticated OAuth user.",
        "responses": {
          "200": {
            "description": "Authenticated user object."
          },
          "401": {
            "description": "Authentication required."
          }
        }
      }
    },
    "/status": {
      "get": {
        "summary": "API status",
        "description": "Returns a simple health check response.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "API is healthy."
          }
        }
      }
    }
  },
  "definitions": {
    "GraphQLRequest": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string",
          "example": "query { recipes { _id title } }"
        },
        "variables": {
          "type": "object"
        }
      }
    }
  }
}